Improving set_available_period_ts_clim

+++ set_available_period_ts_clim.py
@@ -17,24 +17,43 @@
 # -----------------------------------------------------------------------------------------------------
 from CM_atlas import *
 
+# -- Copy the existing one if there is one
+# -- Update!! if the list models is not the same in datasets_setup and datasets_setup_available_period,
+# -- we update it.
+# -- priority to changes in datasets_setup_available_period; try to merge them!!
+
+
 # -- Get the comparison name and build the datasets_setup.py file name
 args = sys.argv
-print 'len(args) = ',len(args)
 if len(args)==1:
     print '--> Provide a comparison name'
 if len(args)==2:
-    print 'Ok'
     comparison = args[1]
     if not comparison[len(comparison)-1]=='/': comparison+='/'
     datasets_setup_file = comparison+'datasets_setup.py'
 
 # -- Create the datasets_setup_period_set.py file
 datasets_setup_available_period_set_file = str.replace(datasets_setup_file,'.py','_available_period_set.py')
-os.system('cp '+datasets_setup_file+' '+datasets_setup_available_period_set_file)
+# -- If datasets_setup_available_period_set_file already exists, we copy it to create a backup
+from datetime import datetime
+delay = datetime.utcnow() - datetime(2015,1,1)
+additionnal = str(delay.microseconds)
+back_up_datasets_setup_available_period_set_file = str.replace(datasets_setup_available_period_set_file, '.py', '_'+additionnal+'.py')
+from datetime import datetime
+if os.path.isfile(datasets_setup_available_period_set_file):
+   os.system('cp '+datasets_setup_available_period_set_file+' '+back_up_datasets_setup_available_period_set_file)
+else:
+   os.system('cp '+datasets_setup_file+' '+datasets_setup_available_period_set_file)
 
 
 # -- Execute the datasets_setup file to get the content
 execfile(datasets_setup_file)
+
+# -- If we already have a datasets_setup_available_period_set_file, we execute it and get Wmodels
+#if os.path.isfile(datasets_setup_available_period_set_file):
+#   execfile(datasets_setup_available_period_set_file)
+
+
 
 # -- Which variable do we use to test the availability of the files?
 period_manager_test_variable = 'tas'
@@ -49,6 +68,7 @@ for dataset_dict in Wmodels_clim:
             frequency_manager_for_diag(dataset_dict, diag='clim')
             get_period_manager(dataset_dict)
             dataset_dict.pop('variable')
+   
 
 # -- TS ---------------------------------------------
 Wmodels_ts = period_for_diag_manager(models, diag='TS')
@@ -62,19 +82,59 @@ for dataset_dict in Wmodels_ts:
             dataset_dict.pop('variable')
 
 
-# -- Append the results to datasets_setup_available_period_set.py       
-thefile = open(datasets_setup_available_period_set_file, 'a')
-print>>thefile, 'Wmodels_ts = ['
+
+thefile = open(datasets_setup_available_period_set_file, 'w')
+print>>thefile, '# -- This file is used as a dataset_setup.py file.'
+print>>thefile, '# -- It is used by main_C-ESM-EP.py if it exists. Remove it (or rename it) if you want to use the period manager again.'
+print>>thefile, '# -- All the modifications that you do here will be used by the C-ESM-EP.'
+print>>thefile, '# -- However, they wont be merged to datasets_setup.py if you want to restart from datasets_setup.py.'
+print>>thefile, ''
+print>>thefile, ''
+print>>thefile, ''
+
+
+# -- Append the results to datasets_setup_available_period_set.py
+list_of_kw = ['project','root','login','model','experiment','simulation']#,'frequency','period','clim_period','ts_period']       
+kw_for_ts   = ['frequency','period','ts_period','diag']
+kw_for_clim = ['frequency','period','clim_period','diag']
+
+print>>thefile, 'Wmodels = ['
+
+item_number = 0
 for item in Wmodels_ts:
-    print>>thefile, item,','
+    print>>thefile,'    dict('
+    for kw in list_of_kw:
+        if kw in item:
+           print>>thefile,'         '+kw+'="'+item[kw]+'",' 
+    for kw in item:
+        if kw not in list_of_kw+kw_for_ts+kw_for_clim:
+           print>>thefile,'         '+kw+'="'+item[kw]+'",'
+    # -- Add period for climatology
+    print>>thefile,'         clim_period = dict('
+    for kw in kw_for_clim:
+        if kw in Wmodels_clim[item_number]:
+           print>>thefile,'                            '+kw+'="'+Wmodels_clim[item_number][kw]+'",'
+    print>>thefile,'                           ),'
+    # -- Add period for the time series
+    print>>thefile,'         ts_period   = dict('
+    for kw in kw_for_ts:
+        if kw in Wmodels_ts[item_number]:
+           print>>thefile,'                            '+kw+'="'+Wmodels_ts[item_number][kw]+'",'
+    print>>thefile,'                           ),'
+    # -- Close the item
+    print>>thefile,'        ),'
+    item_number = item_number + 1
 print>>thefile, ']'
-print>>thefile, ' '
-print>>thefile, ' '
-print>>thefile, ' '
-print>>thefile, 'Wmodels_clim = ['
-for item in Wmodels_clim:
-    print>>thefile, item,','
-print>>thefile, ']'
+
+# -- Append the datasets_setup file
+print>>thefile, ''
+print>>thefile, ''
+print>>thefile, ''
+
+theoldlines = open(datasets_setup_file, 'r').readlines()
+for oldline in theoldlines:
+    print>>thefile, str.replace(oldline,'\n','')
+
 
 thefile.close()
